refactor: source env vars in config module
