+++ src/bot/botClient.js
@@ -1,10 +1,5 @@
-const dotenv = require("dotenv");
 const TelegramBot = require("node-telegram-bot-api");
-
-dotenv.config();
-
-const BOT_TOKEN =
-  process.env.MODE === "stage" ? process.env.STAGE_BOT_TOKEN : process.env.PROD_BOT_TOKEN;
+const { BOT_TOKEN } = require("../config");
 
 const bot = new TelegramBot(BOT_TOKEN, { polling: true });
